# [Fix #385] Disable `Performance/BlockGivenWithExplicitBlock` by default