--- a/changelog/change_disable_block_given.md
+++ b/changelog/change_disable_block_given.md
@@ -1,0 +1,1 @@
+* [#385](https://github.com/rubocop/rubocop-performance/issues/385): Disable `Performance/BlockGivenWithExplicitBlock` by default. ([@earlopain][])