@koic 2024-09-13
If the speed is reversed, shouldn't the behavior be updated to reverse the bad and good cases instead of disabling the cop?