@Earlopain 2024-09-13
I would rather not make that decision. The difference today is smaller than it was when the cop was first introduced and I find it likely to be further optimized in Ruby itself in the future, where the cop can can then again produce worse results.

The results are also only consistently in favor of `block_given?` when YJIT is enabled, otherwise `if block` (without actually passing in a block) is still the fastest variant):
```
               block:  9336250.8 i/s
        block_given?:  8365028.9 i/s - same-ish: difference falls within error
block_given? w/ block:  7620740.6 i/s - 1.23x  slower
      block w/ block:  2334882.5 i/s - 4.00x  slower
```

Overall "it depends". Per https://github.com/rubocop/rubocop-performance/issues/385#issuecomment-1819638731 there is also a difference if the block parameter is defined in the method signature.