@koic 2024-09-13
> The difference today is smaller than it was when the cop was first introduced and I find it likely to be further optimized in Ruby itself in the future

I see. That would make sense.